// Web/src/app/configuration/config-shared/quality-profile-section/quality-profile-store/quality-profile-section-action-type.enum.ts
export enum QualityProfileSectionActionType {

    FetchEntityProfiles = 'FETCH_ENTITY_QUALITY_PROFILES',
    ShowPopup = 'SHOW_POPUP',
    FetchLanguages = 'FETCH_LANGUAGES',
    FetchLanguageProfiles = 'FETCH_LANGUAGE_QUALITY_PROFILE',
    SelectedProfile = 'SELECTED_PROFILE',
    InitialProfile = 'INITIAL_PROFILE'
}